Convert to Ansible playbook
I saw the blog post for this script on Reddit and thought "This is
dying to be an Ansible playbook!"  So I made one.

I tested this on an Ubuntu 16.04 Server VM on my local system and it
appeared to run without any errors.  However, I'm more of a Fedora
guy, so I'm not 100% sure that everything is how it should be.

The playbook runs everything with `sudo` access, so you might want to
tweak it a bit.

Good luck!